openfaas: handling previous lookups on borders
